requirements.txt
numpy==1.22.2
python-dotenv==0.19.2